+++ pkg/engine/globalcontext/store/entry.go
@@ -1,0 +1,6 @@
+package store
+
+type Entry interface {
+	Get() (interface{}, error)
+	Stop()
+}
