+++ pkg/engine/globalcontext/store/store.go
@@ -2,8 +2,6 @@ package store
 
 import (
 	"sync"
-
-	"github.com/pkg/errors"
 )
 
 type Store interface {
@@ -12,70 +10,43 @@ type Store interface {
 	Delete(key string)
 }
 
-type Entry interface {
-	Get() (interface{}, error)
-	Stop()
-}
-
-type invalidentry struct {
-	err error
-}
-
-func (i *invalidentry) Get() (interface{}, error) {
-	return nil, errors.Wrapf(i.err, "failed to create cached context entry")
-}
-
-func (i *invalidentry) Stop() {}
-
-func NewInvalidEntry(err error) Entry {
-	return &invalidentry{
-		err: err,
-	}
-}
-
-type cache struct {
+type store struct {
 	sync.RWMutex
 	store map[string]Entry
 }
 
-func NewStore() Store {
-	return &cache{
+func New() Store {
+	return &store{
 		store: make(map[string]Entry),
 	}
 }
 
-func (l *cache) Set(key string, val Entry) bool {
+func (l *store) Set(key string, val Entry) bool {
 	l.Lock()
 	defer l.Unlock()
-
-	if val, found := l.store[key]; found { // If the key already exists, skip it before replacing it
+	old := l.store[key]
+	// If the key already exists, skip it before replacing it
+	if old != nil {
 		val.Stop()
 	}
-
 	l.store[key] = val
 	_, ok := l.store[key]
 	return ok
 }
 
-func (l *cache) Get(key string) (Entry, bool) {
+func (l *store) Get(key string) (Entry, bool) {
 	l.RLock()
 	defer l.RUnlock()
-
 	entry, ok := l.store[key]
-	if !ok {
-		return nil, false
-	}
-	return entry, true
+	return entry, ok
 }
 
-func (l *cache) Delete(key string) {
+func (l *store) Delete(key string) {
 	l.Lock()
 	defer l.Unlock()
-
-	val, ok := l.store[key]
-	if !ok {
-		return // value already deleted
+	entry := l.store[key]
+	if entry != nil {
+		entry.Stop()
 	}
-	val.Stop()
 	delete(l.store, key)
 }
